warmup + minor updates
school has ended and its time to kick this project back into high gear. Linux compatible open function is still not working, but otherwise I have added a flag so that the program can diffrentiate between a right click menu for the entire directory and a right click for an individual file. I am messing with ways to get the terminal feature working, but that typically requires a subprocess.run call.