revisions, deprecations, and mkdir.
make directory feature added. trashpaths will have to be deprecated because it is not cross platform and linux does not have a local "trash" framework. This is possible but is unrealistic given the scope of my goals for this assignment. Other than that, basic trimming still going and improving of user ergonomics.